+++ .github/actions/metrics-report/src/utils/git.mjs
@@ -1,19 +1,7 @@
 // @ts-check
 
-import path from 'path'
-import { fileURLToPath } from 'url'
-
 import core from '@actions/core'
 import { exec } from '@actions/exec'
-import { getOctokit } from '@actions/github'
-
-const currentDirPath = path.dirname(fileURLToPath(import.meta.url))
-
-export const actionRootPath = path.resolve(currentDirPath, '..')
-export const repoRootPath = path.resolve(currentDirPath, '../../../..')
-
-const githubToken = core.getInput('github_token')
-export const octokit = getOctokit(githubToken)
 
 /**
  * @param {string} branch
