+++ .github/actions/metrics-report/src/utils/github.mjs
@@ -1,0 +1,7 @@
+// @ts-check
+
+import core from '@actions/core'
+import { getOctokit } from '@actions/github'
+
+const githubToken = core.getInput('github_token')
+export const octokit = getOctokit(githubToken)
