+++ .github/actions/metrics-report/src/utils/path.mjs
@@ -1,0 +1,9 @@
+// @ts-check
+
+import path from 'path'
+import { fileURLToPath } from 'url'
+
+const currentDirPath = path.dirname(fileURLToPath(import.meta.url))
+
+export const actionRootPath = path.resolve(currentDirPath, '../..')
+export const repoRootPath = path.resolve(actionRootPath, '../../..')
